+++ kameleoon/client/src/Exception/FeatureExperimentNotFound.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Kameleoon\Exception;
+
+class FeatureExperimentNotFound extends FeatureException
+{
+}